Implementasi final backend BATIM GADAI Web mencakup
helpers, models, migrations, seeders, controllers, views, dan routes.

+++ app/Http/Controllers/GadaiController.php
@@ -181,7 +181,7 @@ class GadaiController extends Controller
     {
         $role = Auth::user()->role;
 
-        if ($role !== 'superadmin' && $gadai->cabang_id !== Auth::user()->cabang_id) {
+        if ($role !== 'superadmin' && (int)$gadai->cabang_id !== (int)Auth::user()->cabang_id) {
             abort(403);
         }
 
@@ -222,7 +222,7 @@ class GadaiController extends Controller
     {
         $role = Auth::user()->role;
 
-        if ($role !== 'superadmin' && $gadai->cabang_id !== Auth::user()->cabang_id) {
+        if ($role !== 'superadmin' && (int)$gadai->cabang_id !== (int)Auth::user()->cabang_id) {
             abort(403);
         }
 
@@ -259,7 +259,7 @@ class GadaiController extends Controller
     {
         $role = Auth::user()->role;
 
-        if ($role !== 'superadmin' && $gadai->cabang_id !== Auth::user()->cabang_id) {
+        if ($role !== 'superadmin' && (int)$gadai->cabang_id !== (int)Auth::user()->cabang_id) {
             abort(403);
         }
